feat: connect back-end model to front-end UI
